+++ UI/Components/Home/LogoDisplay.cs
@@ -161,9 +161,9 @@ namespace PBGame.UI.Components.Home
         /// <summary>
         /// Adjusts the speed of the pulse animation for specified length.
         /// </summary>
-        private void AdjustBeatSpeed(double beatLength)
+        private void AdjustBeatSpeed(float beatLength)
         {
-            pulseAni.Speed = 1000f / (float)beatLength;
+            pulseAni.Speed = 1000f / beatLength;
         }
 
         /// <summary>
